refactor: rename sync worklog to sync time entries
