+++ bin/syncWorklog.js
@@ -3,6 +3,6 @@
 require('../bootstrap');
 const queue = require('../lib/queue');
 
-queue.create('sync worklog').priority('critical').save((err) => {
+queue.create('sync time entries').priority('critical').save((err) => {
   process.exit(!!err);
 });
