+++ package.json
@@ -8,7 +8,7 @@
     "test": "babel-node node_modules/.bin/babel-istanbul cover _mocha && codecov",
     "start": "forever index.js",
     "worker": "forever worker.js",
-    "sync": "node bin/syncWorklog.js",
+    "sync": "node bin/syncTimeEntries.js",
     "watch": "nodemon",
     "knex": "node_modules/.bin/knex"
   },
